Add rule for changing keeper Id (#7)
From the Rule change proposal:

The keeper ID can be automatically changed with the team protocol which was introduced in 2019, but teams could only change the keeper during STOP.
This is an unnecessary limitation. With reduced number of stops, this becomes even more important.

We propose to allow changing the keeper id, as long as the ball is on the opponent's half. This applies to the team protocol as well as to the manual change via the game controller operator.

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>