Emergency stop (Div A) (#11)
From the rule change proposal:

With less stoppages, there are fewer chances to get a robot from the field manually. Even with a timeout, the team would need to wait until the next stoppage.
If something goes really wrong with a teams robot (like damaging itself more and more), the team might want to stop the game to reduce the damage (and any cost due to destroyed parts).

Proposal: A team can ask to stop the game immediately after a grace period of 10s, regardless of the current situation. It will receive a yellow card for this and must take a timeout immediately. If the team is out of timeouts, it is still allowed to remove robots from the field as fast as possible, but can not use any remaining timeout time

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>