Add rule for rule changes during competition (#12)
From the rules proposal:

We as TC/OC/EC try our best to come up with adequate rule changes that will improve the league. But there will always be an uncertainty that can only be clarified by actually running a competition. If we notice, during the competition, that some rule is not working out as expected, we’ll open up the possibilities to adapt the rules:

Between phases of the competition: Round-Robin / knockout / tournament
Only invoked if we find major problems, as a last resort
This is not an “out” for the rules making, to iron out details through regionals, testing, other avenues
We will involve all team leaders into the decision