Remove Penalty Kicks (both Div) (#14)
From the rule proposal:

In 2019, there are two ways to receive a penalty kick:

Every third card led to a penalty kick
If a defender other than the keeper is fully inside the defense area and touches the ball
Penalty kicks have several disadvantages:

They are quite unbalanced (and thus not very interesting)
They lack innovation
They need extra implementation
They require additional time during a match
They interrupt the game
We propose to get rid of penalty kicks completely:

The Multiple Defenders foul will result in a red card. The game will be stopped immediately.
The third active yellow card will be turned into a red card instead. A yellow card is active as long as the counter has not run out, so in order to get a red card, the team must collect three yellow cards within 2 minutes.
During the last team meeting, it was suggested to replace the current penalty kick by the one-on-one procedure that is currently only applied in the shoot-out phase.
This is not as hard of a punishment as the current penalty kick, but it was found a better punishment than the red card.

This PR replaces the penalty kick procedure by the one-on-one procedure. There are no changes in what causes a penalty kick in this PR.