Field drawings update for wall gameplay (#28)
This prepares the field drawings for fast change between playing with and without walls. In the case of playing with walls, the field increases and the goals will be moved accordingly. The defense area will also be increased accordingly, such that the lines of the defense area are still correct.