Challenge flag (Both Div) (#27)
From the rule change proposal:

In American football there is a challenge flag: A team can challenge a decision of the referee
If referees decision was correct, team loses a timeout
If referees decision was incorrect, correct decision is applied
Only one ruling may be challenged at a time.

We consider giving each team about two flags. The exact number is yet to be determined.

The rationale is that we are losing a lot of time discussing referee decisions. It will make teams think twice before discussing a situation.

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>