Tune ball placement procedure (#31)
From the rule proposal:

The ball placement procedure has some corner cases and can be optimized in speed.
There will be some optimizations, that have yet to be determined in detail:
Relax the precision where possible
Avoid ball placement at all, if the ball is already at a good position
Be more strict with placement failures (turn it off faster)

This change implements:

Conditions, when no ball placement is required (ball placement can not be successfully for a team automatically anymore, which would have reset the failure counter before)
Failure counter is not reset on success, but only decremented
Some rational about what has not changed:

Placement tolerance of 0.15m: Increasing this tolerance requires altering other margins as well, like the distance to the field lines for the placement position. As long as the tolerance is not hard to achieve, we should rather keep it as is.