Bugfix: Goals should be invalid, if bot committed foul
Right now, a bot could touch the ball inside the penalty area (committing non-stopping foul) and afterwards kick the ball into the goal. It would be valid, because the foul was before the robot last touched the ball. A 2-second timer will avoid this. 2 seconds is also the grace period for repeating a foul that is still being committed.